refactor: 변수 네이밍 수정 및 패키지 이동
