+++ src/main/java/persistence/sql/column/IdGeneratedStrategy.java
@@ -1,0 +1,6 @@
+package persistence.sql.column;
+
+public interface IdGeneratedStrategy {
+
+    public String getValue();
+}
