+++ src/main/java/persistence/sql/column/MetaDataMapper.java
@@ -2,7 +2,6 @@ package persistence.sql.column;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import persistence.sql.ddl.IdGeneratedStrategy;
 import persistence.sql.dialect.Dialect;
 
 import java.lang.reflect.Field;
@@ -28,12 +27,18 @@ public final class MetaDataMapper {
             nullable = getNullable(isNullable);
         }
         if (field.isAnnotationPresent(Id.class)) {
+            validateGeneratedValue(field);
             GeneratedValue annotation = field.getAnnotation(GeneratedValue.class);
             IdGeneratedStrategy idGeneratedStrategy = dialect.getIdGeneratedStrategy(annotation.strategy());
-
-            return new PkColumn(field, name, columnType);
+            return new PkColumn(name, columnType, idGeneratedStrategy);
         }
         return new JpaColumn(name, columnType, nullable);
+    }
+
+    private static void validateGeneratedValue(Field field) {
+        if (!field.isAnnotationPresent(GeneratedValue.class)) {
+            throw new IllegalArgumentException("[INFO] No @GeneratedValue annotation");
+        }
     }
 
     private static String getNullable(boolean isNullable) {
