--- src/main/java/persistence/sql/ddl/IdGeneratedStrategy.java
@@ -1,4 +1,0 @@
-package persistence.sql.ddl;
-
-public interface IdGeneratedStrategy {
-}
